requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==4.2.9
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
